+++ src/components/TrelloCard/index.tsx
@@ -1,20 +1,20 @@
-import React, { useState } from "react";
-import { TrelloCardProps } from "@/interfaces/trelloCard";
-import { Card } from "antd";
+import React from "react";
 import { Draggable } from "react-beautiful-dnd";
+import { Card } from "antd";
+import { ListsProvider } from "@/context/ListsContext";
+import { TrelloCardProps } from "@/interfaces/trelloCard";
+import ListsHeader from "./ListsHeader";
 import TrelloCardBottomFunc from "./TrelloCardBottomFunc";
-import TrelloCardHeader from "./TrelloCardHeader";
-import { TrelloCardStyled } from "./style";
 import TrelloCardList from "./TrelloCardList";
+import { TrelloCardStyled } from "./style";
 
 const TrelloCard: React.FC<TrelloCardProps> = (props) => {
   const index = props.index;
   const listId = props.quotes.id;
   const listName = props.quotes.name;
-  const [showAddCard, setShowAddCard] = useState(false);
 
   return (
-    <>
+    <ListsProvider>
       <Draggable draggableId={listId} index={index}>
         {(provided, snapshot) => (
           <div ref={provided.innerRef} {...provided.draggableProps}>
@@ -23,13 +23,7 @@ const TrelloCard: React.FC<TrelloCardProps> = (props) => {
               isdragging={snapshot.isDragging.toString()}
             >
               <Card.Meta
-                title={
-                  <TrelloCardHeader
-                    list={props.quotes}
-                    showAddCard={showAddCard}
-                    setShowAddCard={setShowAddCard}
-                  />
-                }
+                title={<ListsHeader listId={listId} />}
                 className="cardTitle"
                 {...provided.dragHandleProps}
                 aria-label={`${listName} quote list`}
@@ -43,18 +37,13 @@ const TrelloCard: React.FC<TrelloCardProps> = (props) => {
                 internalScroll={props.isScrollable}
                 isCombineEnabled={props.isCombineEnabled}
                 useClone={props.useClone}
-                showAddCard={showAddCard}
-                setShowAddCard={setShowAddCard}
               />
-              <TrelloCardBottomFunc
-                showAddCard={showAddCard}
-                setShowAddCard={setShowAddCard}
-              />
+              <TrelloCardBottomFunc />
             </TrelloCardStyled>
           </div>
         )}
       </Draggable>
-    </>
+    </ListsProvider>
   );
 };
