+++ src/components/TrelloCard/style.ts
@@ -47,33 +47,6 @@ export const TrelloCardBottomFuncStyled = styled.div<{ show: string }>`
   }
 `;
 
-export const TrelloCardHeaderStyled = styled.div`
-  display: flex;
-  align-content: center;
-  justify-content: space-between;
-  .button-hover:hover {
-    background-color: var(--black);
-  }
-`;
-
-export const PopoverHeaderStyled = styled.div`
-  position: relative;
-  display: flex;
-  align-content: center;
-  justify-content: center;
-  .title {
-    line-height: 41px;
-  }
-  .popoverCloseIcon {
-    position: absolute;
-    top: 5px;
-    right: 0px;
-  }
-  border-bottom: 1px solid var(--divider-gray);
-`;
-
-export const PopoverContentStyled = styled.div``;
-
 export const TrelloCardAddCss = styled(Card)<{ useadd: string }>`
   padding: 8px;
   min-height: 1px;
